# Parallelise integrators with openMP