openMP directives were introduced in some integrators some time ago by Xiabiao Huang, but were not activated.
They can now be optionally activated.

## Behaviour

In each integrator, the loop over the different particles is parallelised. This occurs if the number of particles is greater than 10 (by default, configurable). Below this number, nothing is changed, so functions like findorbit*x*, findm*xx*, linopt… are not accelerated. on the other hand, when tracking 1000 particles, I could see an acceleration by a factor 2 to 3 on a desktop machine.
At the moment, apart from the nb. particles threshold, everything (like the number of threads) is automatically handled by openMP. Additional functions may be added in the future to get more control.

## Requirements

An openMP-aware compiler is needed. This is what I observed:
#### Linux:
gcc is compatible, nothing to do.
#### MacOS:
For Matlab, we must use the openMP library provided by Matlab: nothing to do,
For python, an openMP library needs to be installed, preferably with a package manager:
`$ sudo port install libomp` or
`$ brew install libomp`
#### Windows:
Nothing to do with Visual Studio Community 2017, to be tested with other compilers.

## Activation

### Matlab
Rebuild the integrators with:
```
>> atmexall -openmp
```
The threshold on the number of particles can be adjusted (to 5 in this example) by using:
```
>> atmexall -openmp -DOMP_PARTICLE_THRESHOLD=5
```
Notice: I experienced crashes when compiling with a Matlab version and running with an older version.

### Python
Set the environment variable `OPENMP` to a non-null value when building pyat. Example for `sh`/`bash`/`zsh`:
```
OPENMP=1 python setup.py build/install/develop
```
You may need to remove previously compiled integrators with:
```
rm -rf build
find at -name "*.pyc" -exec rm '{}' \;
find at -name "*.so" -exec rm '{}' \;
```
The threshold on the number of particles can be adjusted (to 5 in this example) by using:
```
OPENMP=1 OMP_PARTICLE_THRESHOLD=5 python setup.py build/install/develop
```
### Octave
Not yet implemented, in progress…