@matsievskiysv 2021-01-31
Pushed edits for Octave build script.

Tested it using [`testTracking`](https://github.com/atcollab/at/blob/master/atmat/atdemos/tracking/testTracking.m) example with `N=30`:
```octave
tic;testTracking;toc
```
Got 240 and 295 seconds with and without OpenMP respectively.
@lfarv could you maybe suggest a better test case?