@swhite2401 2021-02-01
hello all, 
I did not know this was implemented, could you point to an integrator that has it for me to see the code?

As a general comment on this, openMP is nice but relies on shared memory similar to matlab parfor or pyhton multithreading, as soon as you will want to use several nodes (i.e. in a cluster architecture) you will need something that can deal with distributed memory and openMP cannot. For this, an MPI implementation is needed. 
I also strongly discourage hydrid implementation combining the two. I know this is a much more significant development, but I would go for MPI right away and not spend too much time on this.
The main application for such development is multi-particle tracking and in my opinion it only makes sense if it can be used on a cluster.