@swhite2401 2021-02-01
I took a look and it is in fact very simple, one question though: wouldn't it be simpler to handle this at the atpass.c level?
One could think of splitting the particle array into N array and run the parallelized loop on these.
This way the integrators are left unchanged, and for pyAT there is nothing to do as there the patpass function that implements the same at the python level using multithreading.

Only problem I can see is for the impedance elements that requires a knowledge of all particles, but again here I believe MPI is needed.