@lfarv 2021-02-01
@swhite2401 : As you can see, the implementation is very simple, and none of the integrators is modified in this branch! The only new thing here is the selection of the adequate compile flags for different platforms. An implementation at the `atpass.c` level might be marginally more efficient but requires more work and indeed raises the problem of impedance-like elements.

`patpass` in python has the same problem (impedances), and as far as I could see is slightly less efficient than openMP.

 `parfor` in Matlab is apparently quite different as it does not use shared memory, but spawns several processes (so called "workers"), on a single machine or on several nodes, and circulates the data between them through .mat files. It's more like to MPI, and has a lot more overhead.

I agree that openMP should not be mixed with another parallelisation solution, that's why it's optional and defaults to "non-activated". But the openMP option offers almost for free a solution for any desktop machine, with very little overhead. 

@matsievskiysv : I'll add a test function to the distribution