@swhite2401 2021-02-01
I fully agree, this is very nice and free as you, I was just suggesting further developments.
As far as I can see patpass is not usefull anymore, should it be removed to avoid conflicts?

Ok for me to merge, btw