@lfarv 2021-02-01
I added a test function `trackMultipleParticles` in `demos/tracking`. It tracks 500 particles over 500 tunes (by default) and prints the elapsed time (without and with longitudinal motion). As an example, here are my results (MacOS, quad-core processor):

**Matlab without openMP**
\>> trackMultipleParticles()
Elapsed time is 9.9 seconds.
Elapsed time is 24.7 seconds.

**Matlab with openMP**
\>> trackMultipleParticles()
Elapsed time is 3.2 seconds.
Elapsed time is 7.1 seconds.

**_The gain is approximately a factor 3 !_**

**Octave (native) without openMP**
\>> trackMultipleParticles()
Elapsed time is 9.8 seconds.
Elapsed time is 24.6 seconds.

**Octave (native) with openMP**
\>> trackMultipleParticles()
Elapsed time is 6.2 seconds`                   `<--surprisingly long
Elapsed time is 9.8 seconds.