@lfarv 2021-02-01
Correction: Octave results are equivalent to Matlab results. The problem was that in Octave, clear mex does not ensure that updated files will be loaded. The right results were obtained after reloading Octave.

**Octave (native) with openMP**
\>> trackMultipleParticles()
500 particles, 500 turns, without longitudinal motion: 3.2 s
500 particles, 500 turns,    with longitudinal motion: 7.1 s