@matsievskiysv 2021-02-02
@lfarv enabled OpenMP in Jupyter image