@lfarv 2021-02-02
The choice of the threshold for parallelisation is not obvious: here is a plot showing the ratio between tracking with and without openMP, for 1 to 10 particles (quad-core desktop machine):

![test_opnmp](https://user-images.githubusercontent.com/20038121/106628598-413f1200-657a-11eb-89e3-016cfeb254ea.png)

With a threshold 0 (blue curve, always parallelise), the overhead from spawning the threads increases the elapsed time for n_particles < 5, up to a factor 2 ! In this case, setting OMP_THRESHOLD=4 (red curve) suppresses the problem, and gives some improvement starting with 5 particles (such as in the computation of closed orbit or response matrices).

However, this depends on the machine: here is what I got on another one (Linux server):

![test_opnmp2](https://user-images.githubusercontent.com/20038121/106630872-82382600-657c-11eb-8310-506fe033931a.png)

Here, attempting to parallelise less than 10 particles may give terrible results. However the default threshold of 10 looks safe enough.

To make this check easy, I added a function `test_openmp(threshold)` in` atmat/atdemos/tracking`;
the 1st plot is obtained with

`>> test_openmp(4)
`

and the second with

`>> test_openmp()`