@swhite2401 2021-02-02
Very interesting! I am not sure how openMP works, do you you initialize a new environment every time you enter a passmethod or the same is re-used for the whole tracking?