@lfarv 2021-02-02
@swhite2401 : no idea… For your information, the 1st results comes from my Mac, the second from "grappa". In absolute "grappa" is also much slower, which is disappointing!