@matsievskiysv 2021-02-02
Some considerations:
* Maybe different compilers have different [schedulers](http://jakascorner.com/blog/2016/06/omp-for-scheduling.html) enables by default? I worth investigating which one works best in this case.
* It's probably better changing `defatlt(shared)` to `default(none)`. Intuitively, reducing a number of shared variables should increase performance.
* Particle treshhold should be expressed in particle/thread units.