@swhite2401 2021-02-02
@matsievskiysv thanks for you insight, I was actually documenting myself, and I agree we should limit the number of threads.
From what I have understood so far (correct me if I am wrong) openMP will spawn by default a number of threads = number of CPUs.
The grappa machine has 64 CPUs, so with low particle threshold there are many empty threads that will consume ressources, it could explain some of the issues with speed-up that Laurent observed.
I suggest to add a command that sets the max number of threads<=particle threshold (to be optimized of course)
I think this can be done by simply adding the  num_threads(nthreads) to the existing openMP commands