@matsievskiysv 2021-02-02
It will be more convenient to use [envvars](https://www.ibm.com/support/knowledgecenter/en/SSLTBW_2.2.0/com.ibm.zos.v2r2.cbcux01/ruomprun.htm). This way recompilation or code changes will not be needed.