@matsievskiysv 2021-02-02
BTW, a large number of cores operating on the same memory region would decrease performance due to [cache coherence issues](https://en.wikipedia.org/wiki/Cache_coherence). It's better to split memory into big and continuous chunks. On the first glance [`static`](http://jakascorner.com/blog/2016/06/omp-for-scheduling.html#static) scheduler should work better in this case.