@lfarv 2021-02-02
Concerning the most interesting case of many particles, here is a compilation of my results.

Results are obtained with
`>[t1, t2]=trackMultipleParticles();
`
which tracks 500 particles over 500 turns in the hmba test lattice. I take the average of two runs.
1st line: t1:  no longitudinal motion
2nd line: t2: with longitudinal motion (`atradon`)

|  | quad-core desktop | dual-core laptop | "grappa", 64 CPUs 
|- | - | - | - 
| without OpenMP no rad | 9.9 s | 9.8 s | 12.5 s 
| without OpenMP rad | 24.7 s | 26.3 s | 21.9 s
| |  |  |  
| with OpenMP no rad | 3.2 s | 6.5 s | 3.6 s 
| with OpenMP rad | 7.1 s | 14.0 s | **1.6 s** ??
| |  |  |  
| ratio no rad  | 3.2 | 1.5 | 3.5
| ratio rad  | 3.5 | 1.5 | 13.7
| theoretical ration: n CPUs  | 4 | 2 | 64

Remarks:
- for single-thread, the 3 machines are very similar,
- for multi-thread, the gain on the two Macs is close to the number of cores, which makes sense. For grappa, it seems obvious that the number of threads used is far from the total number of CPUs,
- grappa with openMP is the only example where the computation with radiation on is faster (twice faster) than without !!! To be investigated